Move SparkBaseSuite to spark
