+++ spark/src/test/scala/io/ddf/spark/analytics/AggregationHandlerSuite.scala
@@ -1,6 +1,6 @@
 package io.ddf.spark.analytics
 
-import io.ddf.test.it.SparkBaseSuite
+import io.ddf.spark.SparkBaseSuite
 import io.ddf.test.it.analytics.AggregationHandlerBaseSuite
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
